# Conflicts:
#	src/main/java/com/mh/green2nd/config/InterceptorCorsConfig.java
#	src/main/java/com/mh/green2nd/orders/Orders.java
#	src/main/java/com/mh/green2nd/orders/ordersItem/OrdersItem.java

+++ src/main/java/com/mh/green2nd/menu/MenuDto.java
@@ -28,5 +28,5 @@ public class MenuDto {
     private String menu_imgurl;
     private String menu_bannerimgurl;
 
-    private int menu_price;
+    private double menu_price;
 }
